requirements.txt
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.57.0
oauth2client==4.1.3
oauthlib==3.2.2
pipenv==2022.12.19
python-decouple==3.6
python-dotenv==0.21.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
schedule~=1.1.0